+++ Game/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class GameConfig(AppConfig):
-    name = 'Game'
+    name = 'game'
