+++ Game/templates/create.html
@@ -5,5 +5,3 @@
         <button type="submit" class="save btn btn-default">Save</button>
     </form>
 {% endblock %}
-
-
