+++ Game/templates/view.html
@@ -1,68 +1,37 @@
-{% load leaflet_tags %}
+<!DOCTYPE html>
 <html>
-    <head>
-        ...
-        {% leaflet_js %}
-        {% leaflet_css %}
-        <link rel="stylesheet" href="https://unpkg.com/leaflet-control-geocoder/dist/Control.Geocoder.css" />
-        <script src="https://unpkg.com/leaflet-control-geocoder/dist/Control.Geocoder.js"></script>
-    </head>
-    <body>
-
-        <style>
-            .leaflet-container {  /* all maps */
-                width:  600px;
-                height: 400px;
-            }
-            #specialbigmap {
-                height: 800px;
-            }
-            /* Resize the "display_raw" textbox */
-            .django-leaflet-raw-textarea {
-                width: 100%;
-            }
-        </style>
-
-        <div class="map">
-            {% leaflet_map "yourmap" %}
-        </div>
-
-        <script type="text/javascript">
-        var map;
-            window.addEventListener("map:init", function (e) {
-                var detail = e.detail;
-                // add marker
-                // L.marker([36.71995,-4.4248]).addTo(detail.map);
-
-                // search locations
-                L.tileLayer('https://{s}.tile.osm.org/{z}/{x}/{y}.png', {
-                  attribution: '&copy; <a href="https://osm.org/copyright">OpenStreetMap</a> contributors'
-                }).addTo(detail.map);
-                var geo = L.Control.Geocoder.nominatim({geocodingQueryParams:{countrycodes:"es"}});
-                L.Control.geocoder({geocoder: geo}).addTo(detail.map);
-
-                // click event
-                detail.map.on('click', function(e) {
-                    console.log(e.latlng);
-                    document.getElementById("lat").value = e.latlng.lat;
-                    document.getElementById("lon").value = e.latlng.lng;
-                    L.marker(e.latlng).addTo(detail.map);
-                });
-            }, false);
-
-        </script>
-
-    <!-- example data -->
-        {% if error_message %}<p><strong>{{ error_message }}</strong></p>{% endif %}
-
-        <form action="{% url 'game:save' %}" method="post">
-        {% csrf_token %}
-            <input type="text" name="lat" id="lat">
-            <input type="text" name="lon" id="lon">
-            <input type="submit" value="save">
-        </form>
-
-    </body>
-
-
+  <head>
+    <style>
+       /* Set the size of the div element that contains the map */
+      #map {
+        height: 400px;  /* The height is 400 pixels */
+        width: 100%;  /* The width is the width of the web page */
+       }
+    </style>
+  </head>
+  <body>
+    <h3>My Google Maps Demo</h3>
+    <!--The div element for the map -->
+    <div id="map"></div>
+    <script>
+// Initialize and add the map
+function initMap() {
+  // The location of Uluru
+  var uluru = {lat: -25.344, lng: 131.036};
+  // The map, centered at Uluru
+  var map = new google.maps.Map(
+      document.getElementById('map'), {zoom: 4, center: uluru});
+  // The marker, positioned at Uluru
+  var marker = new google.maps.Marker({position: uluru, map: map});
+}
+    </script>
+    <!--Load the API from the specified URL
+    * The async attribute allows the browser to render the page while the API loads
+    * The key parameter will contain your own API key (which is not needed for this tutorial)
+    * The callback parameter executes the initMap() function
+    -->
+    <script async defer
+    src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAHCaVnPMaUAjAtjpINPxhmHfsr5876u6k&callback=initMap">
+    </script>
+  </body>
 </html>
