--- game/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from . import views
-
-urlpatterns = [
-    path('create/', views.create, name="create"),
-    path('view/', views.view, name="view"),
-]
-
