+++ requirements.txt
@@ -2,7 +2,7 @@ certifi==2019.9.11
 chardet==3.0.4
 dataclasses==0.6
 defusedxml==0.6.0
-Django==2.2.7
+Django==2.2.8
 djongo==1.2.36
 dnspython==1.16.0
 gunicorn==20.0.2
